+++ lib/features/payment/payment_details_page.dart
@@ -1,9 +1,9 @@
 import 'package:collection/collection.dart';
-import 'package:didpay/features/payments/review_request_page.dart';
+import 'package:didpay/features/request/review_request_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:didpay/features/payments/payment_method.dart';
-import 'package:didpay/features/payments/search_payment_methods_page.dart';
+import 'package:didpay/features/payment/payment_method.dart';
+import 'package:didpay/features/payment/search_payment_methods_page.dart';
 import 'package:didpay/l10n/app_localizations.dart';
 import 'package:didpay/shared/theme/grid.dart';
 import 'package:didpay/shared/json_schema_form.dart';
