+++ test/features/payment/payment_details_page_test.dart
@@ -1,7 +1,7 @@
-import 'package:didpay/features/payments/payment_details_page.dart';
+import 'package:didpay/features/payment/payment_details_page.dart';
 import 'package:flutter/material.dart';
-import 'package:didpay/features/payments/payment_method.dart';
-import 'package:didpay/features/payments/search_payment_methods_page.dart';
+import 'package:didpay/features/payment/payment_method.dart';
+import 'package:didpay/features/payment/search_payment_methods_page.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
